Split dashboard status and tab panels
